requirements.txt
anthropic>=0.50.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
